requirements.txt
attrs~=23.2.0
packaging~=23.2
pytest~=7.4.4
Appium-Python-Client~=3.1.1
selenium~=4.16.0